refactor, add logs
